starting on the wizard rewrite
I think I'll merge the first draft of the wizard, then finish
the YAML parsing, then come back to the this. The correct way
to manage this is to make the Wizard wrap around the YAML parser
as much as possible -- just reuse the InstanceBuilders from that.

The we can finish the pluggable version of the wizard.