CV Mode from CLI
OPS (SimStore) CVs have several ways the can get the value for a given
input object: (1) the can evaluate the function; (2) they can find the
result in a memory cache; (3) they can find the result in an on-disk
cache.

The order in which they attempt these define the "mode" of the CV:

* `analysis`: first memory cache, then disk cache, then evaluate
* `production`: first memory cache, the evaluate
* `no-caching`: always evaluate

This PR allows a CLI interface to specify these behaviors. In
particular, this might be useful for users who don't want to save CVs to
disk for cheap CVs, like in toy models.